Creating widget ScoreTable
